requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
psycopg2-binary==2.9.11
gunicorn==23.0.0

